+++ bundles/org.openhab.binding.souliss/src/main/java/org/openhab/binding/souliss/internal/protocol/SoulissCommonCommands.java
@@ -25,7 +25,7 @@ import java.util.Enumeration;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.souliss.SoulissBindingUDPConstants;
+import org.openhab.binding.souliss.internal.SoulissBindingUDPConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
